finally

+++ restore.sh
@@ -1,18 +1,21 @@
 #!/bin/sh
 
-exit_and_log()
+finally()
 {
   echo $@ 1>&2
   logger $@
-  exit $1
+  if [ $1 -eq 0 ]; then
+    sync && reboot
+    exit 0
+  else
+    exit $1
+  fi
 }
 
-cd /backup || exit_and_log 1 No backup directory
+cd /backup || finally 1 No backup directory
 backup=`ls -Art *.tar.gz | tail -n 1`
-if [ -z $backup ]; then exit_and_log 1 No backup files; fi
+if [ -z $backup ]; then finally 1 No backup files; fi
 tar -C / -vxzf $backup
 chmod -R 0755 /etc
 chown -R root:root /etc
-
-exit_and_log 0 Restored $PWD/$backup
-sync && reboot
+finally 0 Restored $PWD/$backup
